requirements.txt
google-generativeai
openai
pandas
spacy
dotenv
anthropic
nltk
langdetect
gensim
scikit-learn